add support for custom socket impl, [NX] more sane reboot payload impl
reboot payload now uses smc to copy the payload to iwram, removing the need to
close sockets and use socket memory as the payload buffer.

the socket impl has changed a lot, and is now generic. iirc i did this for an xbox port
as that doesn't have unix sockets, but i never got around to it.